action listener to threshold in event rate

+++ src/org/processmining/filterd/configurations/FilterdTraceStartEventCategoricalConfig.java
@@ -3,7 +3,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.deckfour.xes.classification.XEventAttributeClassifier;
-import org.deckfour.xes.classification.XEventClass;
 import org.deckfour.xes.classification.XEventClasses;
 import org.deckfour.xes.classification.XEventClassifier;
 import org.deckfour.xes.info.XLogInfo;
@@ -14,9 +13,16 @@ import org.deckfour.xes.model.XTrace;
 import org.processmining.filterd.filters.Filter;
 import org.processmining.filterd.gui.FilterConfigPanelController;
 import org.processmining.filterd.gui.NestedFilterConfigPanelController;
-import org.processmining.filterd.parameters.*;
-import org.processmining.filterd.widgets.*;
+import org.processmining.filterd.parameters.Parameter;
+import org.processmining.filterd.parameters.ParameterMultipleFromSet;
+import org.processmining.filterd.parameters.ParameterOneFromSet;
+import org.processmining.filterd.parameters.ParameterValueFromRange;
+import org.processmining.filterd.parameters.ParameterYesNo;
 import org.processmining.filterd.tools.Toolbox;
+import org.processmining.filterd.widgets.ParameterController;
+import org.processmining.filterd.widgets.ParameterMultipleFromSetController;
+import org.processmining.filterd.widgets.ParameterValueFromRangeController;
+
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
 import javafx.scene.control.Slider;
@@ -149,7 +155,7 @@ public class FilterdTraceStartEventCategoricalConfig extends FilterdAbstractRefe
 								 rate.setChosen("Frequency");
 								 threshold = new ParameterValueFromRange<Integer>(
 											"Frequency threshold", "threshold", 100, null, Integer.TYPE);
-								 threshold.setChosen((Integer)newValue.intValue());
+								 threshold.setChosen(newValue.intValue());
 								 List<String> selection = new ArrayList<String>();
